requirements.txt
einops==0.8.0
numpy==1.24.4
opencv_python==4.8.1.78
Pillow==10.0.0
Pillow==11.0.0
scikit_learn==1.3.0
scipy==1.14.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
